+++ relm4-macros/src/widgets/gen.rs
@@ -1,8 +1,124 @@
-use proc_macro2::TokenStream as TokenStream2;
+use proc_macro2::{Span as Span2, TokenStream as TokenStream2};
 use quote::{quote, quote_spanned, ToTokens};
-use syn::{spanned::Spanned, Error, Ident, Path};
+use syn::{spanned::Spanned, Error, ExprCall, ExprField, ExprPath, Ident, Member, Path, Type};
 
-use super::{Property, PropertyName, PropertyType, Tracker, Widget, WidgetFunc, Generics};
+use super::{Expr, Generics, Property, PropertyName, PropertyType, Tracker, Widget, WidgetFunc};
+
+/// Helper function for the tracker macro.
+fn expr_field_from_expr_call(call_expr: &ExprCall) -> Option<&ExprField> {
+    let first_expr = call_expr.args.iter().next()?;
+    if let Expr::Field(expr_field) = first_expr {
+        Some(expr_field)
+    } else {
+        None
+    }
+}
+
+fn generate_tracker_from_expression(
+    expression: &Expr,
+    model_name: &Type,
+) -> Result<TokenStream2, (Span2, String)> {
+    let error_fn = move |span, msg: &str| {
+        let error_msg =
+                    "Unable to generate tracker function. Please pass a tracker function as the first parameter to the `track!` macro.\n\
+                    Usage: track!(TRACK_CONDITION: bool, FIRST_ARG, SECOND_ARG, ...)";
+        Err((span, format!("{}\nHint:  {}", error_msg, msg)))
+    };
+
+    let unref_expr: &Expr = if let Expr::Reference(expr_ref) = expression {
+        &expr_ref.expr
+    } else {
+        expression
+    };
+
+    let expr_field_opt = match unref_expr {
+        Expr::Call(call_expr) => expr_field_from_expr_call(call_expr),
+        Expr::MethodCall(expr_method_call) => {
+            if let Expr::Field(ref expr_field) = *expr_method_call.receiver {
+                Some(expr_field)
+            } else {
+                None
+            }
+        }
+        _ => None,
+    };
+
+    let expr_field = if let Some(expr_field) = expr_field_opt {
+        expr_field
+    } else {
+        return error_fn(
+            unref_expr.span(),
+            "Couldn't find find a call or method expression.",
+        );
+    };
+
+    let base_is_model = if let Expr::Path(expr_path) = &*expr_field.base {
+        if let Some(ident) = expr_path.path.get_ident() {
+            ident == "model"
+        } else {
+            false
+        }
+    } else {
+        false
+    };
+
+    if !base_is_model {
+        return error_fn(
+            expr_field.base.span(),
+            "Couldn't find a reference to `model`.",
+        );
+    }
+
+    let ident = if let Member::Named(ident) = &expr_field.member {
+        ident.clone()
+    } else {
+        return error_fn(expr_field.member.span(), "Expected a named member");
+    };
+
+    let bool_stream =
+        quote_spanned! { expr_field.span() => model.changed( #model_name::#ident() ) };
+    Ok(bool_stream)
+}
+
+fn component_ident(path: &ExprPath) -> TokenStream2 {
+    if path.path.segments.len() == 1 {
+        let ident = &path.path.segments.first().unwrap().ident;
+        quote_spanned! { path.span() => components.#ident.root_widget() }
+    } else {
+        path.to_token_stream()
+    }
+}
+
+fn component_tokens(expr: &Expr) -> TokenStream2 {
+    match expr {
+        Expr::Call(call) => {
+            if let Expr::Path(path) = &*call.func {
+                if let Some(segs) = path.path.segments.first() {
+                    if segs.ident == "Some" {
+                        if call.args.len() == 1 {
+                            if let Expr::Path(args_path) = call.args.first().unwrap() {
+                                let arg_tokens = component_ident(args_path);
+                                quote_spanned! { path.span() => Some(#arg_tokens) }
+                            } else {
+                                expr.to_token_stream()
+                            }
+                        } else {
+                            expr.to_token_stream()
+                        }
+                    } else {
+                        expr.to_token_stream()
+                    }
+                } else {
+                    expr.to_token_stream()
+                }
+            } else {
+                expr.to_token_stream()
+            }
+        }
+        Expr::Path(path) => component_ident(path),
+        _ => expr.to_token_stream(),
+    }
+}
 
 impl PropertyType {
     fn init_assign_tokens(&self) -> Option<TokenStream2> {
@@ -11,7 +127,14 @@ impl PropertyType {
             PropertyType::Value(lit) => Some(lit.to_token_stream()),
             PropertyType::Watch(tokens) => Some(tokens.to_token_stream()),
             PropertyType::Args(args) => Some(args.to_token_stream()),
-            PropertyType::Track(Tracker { update_fns, .. }) => Some(quote! { #(#update_fns),* }),
+            PropertyType::Track(Tracker {
+                bool_fn,
+                update_fns,
+            }) => Some(if update_fns.is_empty() {
+                quote! { #bool_fn }
+            } else {
+                quote! { #(#update_fns),* }
+            }),
             _ => None,
         }
     }
@@ -31,12 +154,45 @@ impl PropertyType {
         }
     }
 
-    fn track_tokens(&self) -> Option<(TokenStream2, TokenStream2)> {
-        if let PropertyType::Track(tracker) = self {
-            let update_fns = &tracker.update_fns;
-            let update_stream = quote! { #(#update_fns),* };
-            let bool_stream = tracker.bool_fn.to_token_stream();
-            Some((bool_stream, update_stream))
+    fn track_tokens(&self, model_name: &Type) -> Option<(TokenStream2, TokenStream2)> {
+        if let PropertyType::Track(Tracker {
+            update_fns,
+            bool_fn,
+        }) = self
+        {
+            // Only one parameter passed. Try to generate tracker.
+            if update_fns.is_empty() {
+                let bool_stream = match generate_tracker_from_expression(bool_fn, model_name) {
+                    Ok(bool_tokens) => bool_tokens,
+                    Err((span, msg)) => {
+                        return Some((
+                            Error::new(span, &msg).to_compile_error(),
+                            TokenStream2::new(),
+                        ));
+                    }
+                };
+                Some((bool_stream, bool_fn.to_token_stream()))
+            } else {
+                let update_stream = quote! { #(#update_fns),* };
+                let bool_stream = bool_fn.to_token_stream();
+
+                // TODO: Uncomment this and add a warning once proc-macro warning are stable
+                /*if update_fns.len() == 1 {
+                    if let Ok(auto_bool_stream) =
+                        generate_tracker_from_expression(&update_fns[0], model_name)
+                    {
+                        if auto_bool_stream.to_string() == bool_stream.to_string() {
+                            let error_msg = "Consider removing the first parameter because the macro would generate the same code.\n";
+                            /*println!(Some((
+                                Error::new(bool_fn.span(), error_msg).to_compile_error(),
+                                update_stream,
+                            ));*/
+                        }
+                    }
+                }*/
+
+                Some((bool_stream, update_stream))
+            }
         } else {
             None
         }
@@ -53,7 +209,7 @@ impl PropertyType {
     fn component_tokens(&self) -> Option<TokenStream2> {
         match self {
             PropertyType::Widget(widget) => Some(widget.widget_assignment()),
-            PropertyType::Component(expr) => Some(quote_spanned! { expr.span() => #expr }),
+            PropertyType::Component(expr) => Some(component_tokens(expr)),
             _ => None,
         }
     }
@@ -347,12 +503,12 @@ impl Widget {
         stream
     }
 
-    pub fn track_stream(&self) -> TokenStream2 {
+    pub fn track_stream(&self, ty: &Type) -> TokenStream2 {
         let w_name = &self.name;
         let mut stream = TokenStream2::new();
 
         for prop in &self.properties.properties {
-            let p_assign_opt = prop.ty.track_tokens();
+            let p_assign_opt = prop.ty.track_tokens(ty);
             if let Some((bool_stream, update_stream)) = p_assign_opt {
                 let p_name = &prop.name;
                 let p_span = p_name.span().unwrap().into();